+++ rtp/rtp_test.go
@@ -264,6 +264,23 @@ func Test_RTPTags(t *testing.T) {
 	}, []int{0, 1, 2, 3, 4, 5, 6, 7}, []int{1, 2, 3, 4, 5})
 	rtp.Root.AddChild("fg", fg)
 
+	rtp.AddHitRateNode("special", func(node *HitRateNode, pr *sgc7game.PlayResult) bool {
+		gp, isok := pr.CurGameModParams.(*gameparams)
+		if isok && gp != nil {
+			if gp.isSpecial && node.TagName == "special" {
+				node.TriggerNums++
+
+				for range pr.Results {
+					node.TotalNums++
+				}
+
+				return true
+			}
+		}
+
+		return false
+	})
+
 	rtp.CalcRTP()
 
 	pr0 := &sgc7game.PlayResult{
