+++ unittestdata/rtptesttagok.csv
@@ -38,3 +38,8 @@ fg,special,7,300,0,0,0,0,0,0
 fg,special,,300,,,,,,300
 fg,,,300,,,,,,300
 ,,,300,,,,,,900
+
+
+
+name,hitrate,average
+special,0.3333333333333333,3
