+++ ver/version.go
@@ -1,4 +1,4 @@
 package sgc7ver
 
 // Version -
-const Version = "v0.2.28"
+const Version = "v0.2.29"
